backend/api/v2/requirements.txt
"""
Requirements for GraphQL API v2 Implementation
Install these dependencies to enable GraphQL functionality
"""

# GraphQL Core Dependencies
strawberry-graphql==0.206.0      # Main GraphQL library for Python
strawberry-graphql[fastapi]==0.206.0  # FastAPI integration

# Additional Dependencies for GraphQL Features
graphql-core==3.2.3             # GraphQL core library
graphql-relay==3.2.0             # Relay specification support

# WebSocket Support for Subscriptions  
websockets==11.0.3               # WebSocket support
uvicorn[standard]==0.23.2        # ASGI server with WebSocket support

# JWT Token Support (if not already installed)
PyJWT==2.8.0                     # JWT token handling
cryptography==41.0.7             # JWT cryptographic support

# Development Tools
graphql-playground==1.0.2        # GraphQL Playground (alternative to GraphiQL)

# Optional: Advanced GraphQL Features
graphql-ws==0.4.4                # GraphQL WebSocket protocol
dataloader==2.0.0                # DataLoader pattern for efficient queries
